59 upstream commits since 36e28e7. Three files conflicted; everything
else merged on its own, including the address-widening surface, uTP and
the NAT rendezvous work.

src/LibSocket.h and the second hunk of src/BaseClient.cpp were
adjacency, not disagreement: each side had added something at the same
point (uTP accessors and EnableTcpNoDelay; RetryNextAddressFamily and
SendSharedFilesRequest). Both sides are kept in each.

src/ClientList.h took upstream's side on the browse. aa34fe3 gave the
browse an owner in CBrowseManager and deleted CClientList's list along
with every call site, so after the merge the only references left were
the declarations themselves -- keeping them would have kept dead code.
The callback-request throttle keeps the fork's signature, which takes a
CNetworkAddress and counts against the peer's rate-limit scope so an
IPv6 peer is throttled at /64 rather than per /128.

src/BaseClient.cpp is the one real decision. Upstream extracted the
contact preconditions into CheckContactPreconditions(), which is a
better shape and now has two callers -- but its version reintroduced the
defect the widening had fixed: a native IPv6 peer has no 32-bit form, so
`if (!uClientIP) return Contacting;` waved it past both the ipfilter and
the ban check. The extraction is kept and the widening moved inside it,
so the address the peer will actually be dialled at is what gets
filtered, and absence is distinguished from "no IPv4 form".

po/ was regenerated with scripts/update-po.sh rather than resolved by
hand; a second pass changes nothing.

Verified: build clean, 69/69 tests pass -- 63 before the merge, plus the
six upstream brought, including BrowseLifecycleTest and BrowseStoreTest,
which cover the manager this merge adopted.

+++ src/SearchFile.h
@@ -197,7 +197,7 @@ public:
 		uint32_t m_ip;
 		uint16_t m_port;
 		uint32_t m_serverIP;
-		uint32_t m_serverPort;
+		uint16_t m_serverPort;
 	};
 
 	void AddClient(const ClientStruct &client);
